requirements.txt
gradio==4.19.2
huggingface-hub==0.21.3
python-dotenv==1.0.0
requests==2.31.0
PyPDF2>=3.0.0
